test(android): verify Python golden parity on 96 captured frames
- ParityTest replays frames.json + golden.json (shared with the Swift
  parity test via gradle resources.srcDir → Packages/WhoopProtocol/Tests
  /Resources). Every parsed dict matches Python's reference output.
- ParsedValueSerializer (kotlinx.serialization) reads bare JSON scalars
  with the same Bool > Int > Double > String precedence Swift uses,
  including Int-prefer-for-integral-numbers so "3644.0" decodes to IntV.
- u32 reads now widen to Long (unsigned) and choose IntV vs DoubleV
  via intOrDoubleV(). Without this, record_hdr / unknown_hdr / clock
  values past 2^31 wrapped to negative Int and broke golden equality.
- Touched callers: Interpreter.parseFrame static-field loop,
  PostHooks.historical_data, GET_CLOCK, METADATA unix/unk0/trim_cursor.
- Added ParsedValue.longValue and switched HistoricalMeta.End to use it
  so trim_cursor > 2^31 stays comparable across IntV / DoubleV.

61/61 tests pass: 2 parity + 12 framing + 13 schema + 10 interpreter
+ 11 posthooks + 8 streams + 5 historical-meta.

+++ android/protocol/src/main/kotlin/com/whoop/protocol/HistoricalMeta.kt
@@ -14,8 +14,8 @@ fun classifyHistoricalMeta(p: ParsedFrame): HistoricalMeta {
         metaName.startsWith("HISTORY_START") -> HistoricalMeta.Start
         metaName.startsWith("HISTORY_COMPLETE") -> HistoricalMeta.Complete
         metaName.startsWith("HISTORY_END") -> {
-            val unix = p.parsed["unix"]?.intValue ?: return HistoricalMeta.Other
-            val trim = p.parsed["trim_cursor"]?.intValue ?: return HistoricalMeta.Other
+            val unix = p.parsed["unix"]?.longValue ?: return HistoricalMeta.Other
+            val trim = p.parsed["trim_cursor"]?.longValue ?: return HistoricalMeta.Other
             HistoricalMeta.End(unix = unix.toUInt(), trim = trim.toUInt())
         }
         else -> HistoricalMeta.Other
